+++ src/ast/mod.rs
@@ -3,11 +3,13 @@ pub mod ext;
 pub mod query_visitor;
 /// Utilities visiting GraphQL AST trees
 pub mod schema_visitor;
+pub mod type_info;
 pub mod type_info_query_visitor;
 pub mod utils;
 
 pub use self::ext::*;
 pub use self::query_visitor::*;
 pub use self::schema_visitor::*;
+pub use self::type_info::*;
 pub use self::type_info_query_visitor::*;
 pub use self::utils::*;
